Remove redundant comparison to arr.Indexof(arr[i])
What does indexOf(n) do? It returns the INDEX of the given VALUE (n in this case) for the array your asking about.

Imagine we have an array A that is { 11, 12, 13, 14 }.

Now, if we iterate over the array with a for-loop...

for (let i = 0; i < A.length; i++) {
    if (A.indexOf(A[i]) { // Do something... }
}

This is very similar to what you had.

But let's break it down, what are you actually asking here? It's easy to see the uselessness of this when you simulate the values `i` will have over the course of the for-loop. Let's start at the beginning...

i = 0 so...
A.indexOf(A[0]) which can be evaluated further to...
A.indexOf(11) because 11 is the value at the 0th index of A

At this point it should become obvious that the INDEX of 11 for the array A will be 0! So we are asking a question we already know the answer to, so it is completely redundant work.