Simplify if-statement
I think this is the best way to express what we want to happen in this function. Instead of checking to see if each value in `args` is also in `arr` (whether through directly accessing them, as you did, or through a nested loop), we instead flip this around and ask, "Does the current value of `arr` have an index in `args`?" If it does, we skip, and if it doesn't, we push it.